move wallet to latest rpc patterns
